requirements.txt
Appium-Python-Client
selenium
pytest
invoke
pytest-html-reporter